feat: implements tracking and api
Uses SORT for tracking for it's good performance and relative simplicity via a
Tracker abstraction. Should make it rather straight forward to use other
tracking algorithms in the future.

Implements a /tracking/track endpoint for posting lists of frames consisting of
detections, which returns back a list of tracked objects.