fix: remove_job test use copy of jobs
Fixes unintended behaviour in the test, where the modified project could
change itself since its the same memory address. Fixes by using a copy
of the jobs list in order to check if they are removed from a project.
This is significant if job were to use list instead of set to store
jobs.