refactor: change jobs in project from set to list
Refactors project class to hold a list of jobs instead of a set. This
makes it easier to use with API since the job objects can be sent
together with the get request for a project. Changes api call for
project to have json with project data variables, along with a list of
the jobs. Also modifies tests to check for the extra data from the
project endpoint. Renames `_jobs` variable in project to `jobs`.