feat(api): expand project variables
Adds a location string variable in project class. Allows the user to mark where the physical location of the project is located. Modifies API to optionally work with location. If the location string is not set during creation, it will be `None` as this is an Optional.

Modifies project in a major way, changing and renaming _jobs variable to be a list instead of a set. This was done to be able to work better with the API endpoint. Because of this it also modifies `project_get` endpoint to also include the list of jobs in the request body. That way the front end only needs a single API call to populate the UI with data.

Added tests to check location functionality. Modifies some tests due to them checking raw json from API. When adding location they were throwing errors due to there were additional variables.