feat: add basic job processing
Very ugly implementaton that has to import a lot from sqlalchemy in
order to create a session to get a job. Currently it will block the
entire api when processing. This is due to that it holds the mappers for
the entire duration of processing. Which would make new API calls throw errors
for the duration. We need a better way to create repositories.

Noticed that probability for the abbor-small video was 0, possibly a bug
in the code to calculate label.

But it "just works" now.