feat: rewrite api testing of start and stop
Changes start and pause to be of HTTP staus code 202. Which tells the
caller that the request has been accepted, and is queued for procesisng.
This is because of scheduler is living in its own thread, and needs time
to process the request.

This also decouples the API from the implementation of job status in the
model. Which simplify the tests a lot.