build: adds new targets to makefile
Makes the makefile the one stop shop for anything related to building
this code. It will set up virtual environment, install dependencies and
let you easily run the software.

This removes any fear of polluting your system with pip packages by
mistake. It also allows you to rename the virtual environment to
anything that suits you. It is not recommended to rename the virtual
environment to `venv`. I have tested and it works, but produces
warnings.

updates .gitignore to not track files generated by make.