Stat calc (#75)
* docs: fix typo in docstring

* feat: job stat property

* feat: `Job` and `JobBare` gets `JobStat` member

JobStats is the pydantic version for model.Job.stats.

Also converts labels from `int` to `str` for `JobStat`. A bug surfaced
from this where `convert_labels` happens twice. We simply ignore and
return if every key is a valid `str` that cannot be cast to an int.

* test: all tests for new job statistic generation

Tests model.Job.stats property, as well as the (de)serialization with
pydantic.

* refactor: moves api helper functions to api/utils