fix: optimized tracking parameters
After testing every permutation of `max_age` and `min_hits` across every
dataset created during our bachelor project, these are the parameters
I've settled on for the best average total sum accuracy.

The tests were done by first using our model to detect objects in our
annotated data sets. These were then serialized into JSON and saved. The
detection then go deserialized and input into SORT. these were then ran
through SORT using all permutations of `max_age` and `min_hits` in a
range between 1 and 25.

`max_age=8`: This allows a hypothesis to exist for 8/25th of a second
before it is dropped. This means an object can be out of frame, obscured
or just missing for 8 frames.

`min_age=1`: This allows an object to have existed for 1/25th of a
second. This seems very low, and somewhat silly, however best results
were achieved with this as low as possible.

I think these results prove that our detection model isn't ideal. It
seems to have a fair bit of false positives, and this is the best way to
compensate such that the total sum is as accurate as possible.

Other interesting results were that individual fish are tracked best
when `max_age` and `min_age` is low. For packs of fish `max_age` close
to 20 and `min_hits` close to 10 seems to provide the best results.

In an ideal world, I would argue that `min_age` should be set closer to
a full second. Meaning an object must exist in the frame for a second to
be counted. `max_age` should probably also be set higher ideally. It
will allow more occlusion, which after visual inspection happens a lot,
but it might bring the side effect of more miss matching.

There is still another parameter to play with, which is `iou`,
intersection over union. Given that `max_age` is high, I don't see any
point of trying to mess with it's default value of 0.3. Increasing this
parameter means the detections has to be closer together, which
`max_age` probably works against. It might be worth looking at in the
future.

closes #62