fix: calculate timestamp based on video length
First video will parse the timestamp in `from_path`. This is then
collected and the correct timestamp for the next video is calculated
based on the length of the previous video.

Also add support for multiple timestamps This will make let a user add
videos with different timestamp to the same job, for instance two
different days, and timestamps gets calculated correctly.